requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
clickhouse-connect==0.8.18
redis[async]==5.0.8
aioredis==2.0.1
celery==5.4.0
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b
websockets==12.0
aiohttp==3.10.5
orjson==3.10.7
python-dotenv==1.0.1
structlog==24.4.0
